// frontend/src/api/generated/models/collection-event-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * linkweave-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ChangeKind } from './change-kind';
import {
    ChangeKindFromJSON,
    ChangeKindFromJSONTyped,
    ChangeKindToJSON,
    ChangeKindToJSONTyped,
} from './change-kind';

/**
 * 
 * @export
 * @interface CollectionEventJson
 */
export interface CollectionEventJson {
    /**
     * 
     * @type {string}
     * @memberof CollectionEventJson
     */
    collectionId: string;
    /**
     * 
     * @type {string}
     * @memberof CollectionEventJson
     */
    bookmarkId?: string;
    /**
     * 
     * @type {ChangeKind}
     * @memberof CollectionEventJson
     */
    kind: ChangeKind;
    /**
     * 
     * @type {string}
     * @memberof CollectionEventJson
     */
    originClientId?: string;
    /**
     * 
     * @type {string}
     * @memberof CollectionEventJson
     */
    actorName?: string;
}



/**
 * Check if a given object implements the CollectionEventJson interface.
 */
export function instanceOfCollectionEventJson(value: object): value is CollectionEventJson {
    if (!('collectionId' in value) || value['collectionId'] === undefined) return false;
    if (!('kind' in value) || value['kind'] === undefined) return false;
    return true;
}

export function CollectionEventJsonFromJSON(json: any): CollectionEventJson {
    return CollectionEventJsonFromJSONTyped(json, false);
}

export function CollectionEventJsonFromJSONTyped(json: any, ignoreDiscriminator: boolean): CollectionEventJson {
    if (json == null) {
        return json;
    }
    return {
        
        'collectionId': json['collectionId'],
        'bookmarkId': json['bookmarkId'] == null ? undefined : json['bookmarkId'],
        'kind': ChangeKindFromJSON(json['kind']),
        'originClientId': json['originClientId'] == null ? undefined : json['originClientId'],
        'actorName': json['actorName'] == null ? undefined : json['actorName'],
    };
}

export function CollectionEventJsonToJSON(json: any): CollectionEventJson {
    return CollectionEventJsonToJSONTyped(json, false);
}

export function CollectionEventJsonToJSONTyped(value?: CollectionEventJson | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'collectionId': value['collectionId'],
        'bookmarkId': value['bookmarkId'],
        'kind': ChangeKindToJSON(value['kind']),
        'originClientId': value['originClientId'],
        'actorName': value['actorName'],
    };
}